refactor: refactor code by FSD
