+++ src/shared/__tests__/Registration.test.tsx
@@ -1,7 +1,7 @@
 import store from '@shared/redux';
 import '@testing-library/jest-dom';
 import { act, fireEvent, render, screen } from '@testing-library/react';
-import RegistrationComponent from '@widgets/authorization/Registration/ui/Registration';
+import { Registration } from '@widgets/authorization';
 import { Provider } from 'react-redux';
 import userEvent from 'user-event';
 
@@ -9,7 +9,7 @@ describe('Registration component', () => {
   beforeEach(() => {
     render(
       <Provider store={store}>
-        <RegistrationComponent />
+        <Registration />
       </Provider>,
     );
   });
