--- src/widgets/authorization/IconPassword/ui/index.ts
@@ -1,3 +1,0 @@
-import IconPassword from './IconPassword';
-
-export { IconPassword };
