--- src/widgets/authorization/Login/ui/index.ts
@@ -1,3 +1,0 @@
-import Login from './Login';
-
-export { Login };
