--- src/widgets/authorization/Logout/ui/index.ts
@@ -1,3 +1,0 @@
-import Logout from './Logout';
-
-export { Logout };
