--- src/widgets/authorization/Registration/ui/index.ts
@@ -1,3 +1,0 @@
-import Registration from './Registration';
-
-export { Registration };
