--- src/widgets/authorization/TextFieldHint/ui/index.ts
@@ -1,3 +1,0 @@
-import TextFieldHint from './TextFieldHint';
-
-export { TextFieldHint };
