+++ src/widgets/authorization/index.ts
