+++ src/widgets/authorization/ui/Logout.tsx
@@ -1,5 +1,5 @@
 import { Button } from '@mui/material';
-import { logout } from '../../model/firebase';
+import { logout } from '../model/firebase';
 import { useAppDispatch } from '@shared/hooks/hook';
 import { showAlert } from '@shared/redux/slices/alertSlice';
 import { AlertStyle } from '@widgets/alert/model/Alert.model';
