+++ src/widgets/authorization/ui/index.ts
@@ -1,0 +1,6 @@
+import Login from './Login';
+import Logout from './Logout';
+import Registration from './Registration';
+import VisibilityButton from './VisibilityButton';
+
+export { Login, Logout, Registration, VisibilityButton };
