+++ tsconfig.json
@@ -28,6 +28,12 @@
       "@/*": ["./*"]
     }
   },
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
+  "include": [
+    "next-env.d.ts",
+    "**/*.ts",
+    "**/*.tsx",
+    ".next/types/**/*.ts",
+    "src/widgets/authorization/ui/VisibilityButton.tsx"
+  ],
   "exclude": ["node_modules"]
 }
